--- pkg/config/interfaces.go
@@ -1,14 +1,0 @@
-package config
-
-// Loader defines methods for loading configuration.
-type Loader interface {
-	Load() (*Config, error)
-	SetConfigFilePath(path string) error
-}
-
-// Exporter defines methods for exporting configuration.
-type Exporter interface {
-	ToYAML(filePath string) error
-	ToJSON(filePath string) error
-	ToEnv(filePath string) error
-}
